refactor: 데이터 레이어 패키지 구조 개선 및 AuthDataSource 구현
- 클린 아키텍처 원칙에 따른 data 모듈 패키지 구조 재설계
  - datasource (Interface)와 remote/local (Implementation) 분리
  - 확장성을 고려하여 remote 하위에 firestore, auth, storage 폴더 세분화
- 인증(Auth) 로직 고도화
  - AuthDataSource 인터페이스 정의 (Login, SignUp, Logout, Delete 등)
  - FirebaseAuthDataSourceImpl 구현 완료 (DataResourceResult 기반 예외 처리)
- 로컬 설정 및 앱 상태 관리 준비
  - AppSettings 도메인 모델 및 SettingsEntity 추가
  - SettingsMapper를 통한 데이터 변환 로직 구현 (자동 로그인 상태 유지 대비)
- 공통 처리 로직 개선
  - DataResourceResult 구조화 (Success, Failure, Loading)
  - 매핑 편의를 위한 mapData 확장 함수 추가

+++ data/src/main/java/com/subin/leafy/data/model/dto/CommunityPostDto.kt
@@ -18,7 +18,7 @@ data class CommunityPostDto(
 
     // 요약 정보들
     val teaType: String? = null,
-    val rating: Float = 0f,
+    val rating: Int = 0,
     val tags: List<String> = emptyList(),
     val brewingSummary: String? = null,
